requirements.txt
click==8.1.7
PyGithub==2.1.1
python-gitlab==4.1.1
google-cloud-aiplatform==1.35.0
